#107 비회원의 약속 삭제하기 개선

+++ src/main/java/com/example/gatherplan/appointment/service/impl/TempAppointmentServiceImpl.java
@@ -130,9 +130,21 @@ public class TempAppointmentServiceImpl implements TempAppointmentService {
     @Override
     @Transactional
     public void deleteAppointment(DeleteTempAppointmentReqDto reqDto) {
-        Appointment appointment = customAppointmentRepository
-                .findByAppointmentCodeAndTempUserInfoAndUserRole(reqDto.getAppointmentCode(), reqDto.getTempUserInfo().getNickname(), reqDto.getTempUserInfo().getPassword(), UserRole.HOST)
+        Appointment appointment = appointmentRepository.findByAppointmentCode(reqDto.getAppointmentCode())
                 .orElseThrow(() -> new AppointmentException(ErrorCode.NOT_FOUND_APPOINTMENT));
+
+        if (AppointmentState.CONFIRMED.equals(appointment.getAppointmentState())) {
+            throw new AppointmentException(ErrorCode.APPOINTMENT_ALREADY_CONFIRMED);
+        }
+
+        UserAppointmentMapping userAppointmentMapping =
+                userAppointmentMappingRepository.findByAppointmentCodeAndNicknameAndTempPassword(
+                                reqDto.getAppointmentCode(), reqDto.getTempUserInfo().getNickname(), reqDto.getTempUserInfo().getPassword())
+                        .orElseThrow(() -> new AppointmentException(ErrorCode.NOT_FOUND_USER_APPOINTMENT_MAPPING));
+
+        if (!UserRole.HOST.equals(userAppointmentMapping.getUserRole())) {
+            throw new AppointmentException(ErrorCode.USER_NOT_HOST);
+        }
 
         userAppointmentMappingRepository.deleteAllByAppointmentCode(reqDto.getAppointmentCode());
         appointmentRepository.deleteById(appointment.getId());
